server/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.3
blis==0.7.4
Bottleneck==1.3.2
CacheControl==0.12.6
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==6.7
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
cymem==2.0.5
distlib==0.3.0
distro==1.4.0
fastai==1.0.58
fastcore==1.3.19
fastprogress==1.0.0
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==0.24
Jinja2==2.11.2
joblib==1.0.1
kiwisolver==1.3.1
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.3.4
msgpack==0.6.2
murmurhash==1.0.5
numexpr==2.7.2
nvidia-ml-py3==7.352.0
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
progress==1.5
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
scikit-learn==0.24.1
six==1.14.0
soupsieve==2.2
spacy==2.3.5
SQLAlchemy==1.1.11
sqlalchemy-cockroachdb==1.3.2
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.25.8
wasabi==0.8.2
webencodings==0.5.1
Werkzeug==1.0.1
